requirements.txt
h5py==3.2.1
numpy==1.20.3
pandas==1.2.4
Pillow==9.0.1
scikit_learn==1.0.2
six==1.16.0
tensorflow==2.3.1
torch==1.10.2
torchvision==0.8.1
tqdm==4.42.1
